refactor(common, table): 테이블, common 멀티 모듈 구성
