+++ common/build.gradle
@@ -1,0 +1,3 @@
+dependencies {
+    compile 'org.springframework.boot:spring-boot-starter-data-jpa'
+}
