+++ common/src/main/java/kichenpos/common/domain/Validator.java
@@ -1,0 +1,11 @@
+package kichenpos.common.domain;
+
+public interface Validator<T> {
+
+    static <T> Validator<T> fake() {
+        return target -> {
+        };
+    }
+
+    void validate(T target);
+}
