+++ common/src/main/java/kichenpos/common/exception/InvalidStatusException.java
@@ -1,0 +1,8 @@
+package kichenpos.common.exception;
+
+public final class InvalidStatusException extends RuntimeException {
+
+    public InvalidStatusException(String message) {
+        super(message);
+    }
+}
