+++ common/src/main/java/kichenpos/common/exception/NotFoundException.java
@@ -1,0 +1,8 @@
+package kichenpos.common.exception;
+
+public final class NotFoundException extends RuntimeException {
+
+    public NotFoundException(String message) {
+        super(message);
+    }
+}
