--- src/main/java/kitchenpos/common/domain/Validator.java
@@ -1,11 +1,0 @@
-package kitchenpos.common.domain;
-
-public interface Validator<T> {
-
-    static <T> Validator<T> fake() {
-        return target -> {
-        };
-    }
-
-    void validate(T target);
-}
