--- src/main/java/kitchenpos/common/exception/InvalidStatusException.java
@@ -1,8 +1,0 @@
-package kitchenpos.common.exception;
-
-public final class InvalidStatusException extends RuntimeException {
-
-    public InvalidStatusException(String message) {
-        super(message);
-    }
-}
