--- src/main/java/kitchenpos/common/exception/NotFoundException.java
@@ -1,8 +1,0 @@
-package kitchenpos.common.exception;
-
-public final class NotFoundException extends RuntimeException {
-
-    public NotFoundException(String message) {
-        super(message);
-    }
-}
