--- src/main/java/kitchenpos/table/domain/CustomerStatus.java
@@ -1,14 +1,0 @@
-package kitchenpos.table.domain;
-
-enum CustomerStatus {
-
-    EMPTY, SEATED, ORDERED, FINISH;
-
-    boolean isEmpty() {
-        return this == EMPTY;
-    }
-
-    boolean isOrdered() {
-        return this == ORDERED;
-    }
-}
