+++ table/http/table-groups.http
@@ -1,0 +1,19 @@
+###
+POST {{host}}/api/table-groups
+Content-Type: application/json
+
+{
+  "orderTables": [
+    {
+      "id": 1
+    },
+    {
+      "id": 2
+    }
+  ]
+}
+
+###
+DELETE {{host}}/api/table-groups/1
+
+###
