+++ table/http/tables.http
@@ -1,0 +1,45 @@
+###
+POST {{host}}/api/tables
+Content-Type: application/json
+
+{
+  "numberOfGuests": 0,
+  "empty": true
+}
+
+###
+GET {{host}}/api/tables
+
+###
+PUT {{host}}/api/tables/1/empty
+Content-Type: application/json
+
+{
+  "empty": false
+}
+
+###
+PUT {{host}}/api/tables/1/number-of-guests
+Content-Type: application/json
+
+{
+  "numberOfGuests": 4
+}
+
+###
+PUT {{host}}/api/tables/1/empty
+Content-Type: application/json
+
+{
+  "empty": true
+}
+
+###
+PUT {{host}}/api/tables/2/empty
+Content-Type: application/json
+
+{
+  "empty": true
+}
+
+###
