+++ table/src/main/java/kichenpos/table/group/domain/OrderTables.java
@@ -1,0 +1,80 @@
+package kichenpos.table.group.domain;
+
+import java.util.Collections;
+import java.util.List;
+import java.util.Objects;
+import javax.persistence.CascadeType;
+import javax.persistence.Embeddable;
+import javax.persistence.OneToMany;
+import kichenpos.table.table.domain.OrderTable;
+import org.springframework.util.Assert;
+
+@Embeddable
+class OrderTables {
+
+    @OneToMany(cascade = {CascadeType.PERSIST, CascadeType.MERGE}, mappedBy = "tableGroup")
+    private List<OrderTable> tables;
+
+    protected OrderTables() {
+    }
+
+    private OrderTables(List<OrderTable> tables) {
+        Assert.notNull(tables, "주문 테이블 리스트는 필수입니다.");
+        Assert.noNullElements(tables,
+            () -> String.format("주문 테이블 리스트(%s)에 null이 포함될 수 없습니다.", tables));
+        this.tables = tables;
+    }
+
+    static OrderTables from(List<OrderTable> tables) {
+        return new OrderTables(tables);
+    }
+
+    List<OrderTable> list() {
+        return Collections.unmodifiableList(tables);
+    }
+
+    boolean notHaveGroupAndEmpty() {
+        return tables.stream().allMatch(OrderTable::notHaveGroupAndEmpty);
+    }
+
+    boolean anyOrdered() {
+        return tables.stream().anyMatch(OrderTable::isOrdered);
+    }
+
+    int size() {
+        return tables.size();
+    }
+
+    void changeGroup(TableGroup tableGroup) {
+        Assert.notNull(tableGroup, "변경하려는 그룹은 필수입니다.");
+        tables.forEach(table -> table.changeGroup(tableGroup));
+    }
+
+    void ungroup() {
+        tables.forEach(OrderTable::ungroup);
+    }
+
+    @Override
+    public int hashCode() {
+        return Objects.hash(tables);
+    }
+
+    @Override
+    public boolean equals(Object o) {
+        if (this == o) {
+            return true;
+        }
+        if (o == null || getClass() != o.getClass()) {
+            return false;
+        }
+        OrderTables that = (OrderTables) o;
+        return Objects.equals(tables, that.tables);
+    }
+
+    @Override
+    public String toString() {
+        return "OrderTables{" +
+            "tables=" + tables +
+            '}';
+    }
+}
