+++ table/src/main/java/kichenpos/table/group/domain/TableGroupRepository.java
@@ -1,0 +1,12 @@
+package kichenpos.table.group.domain;
+
+import kichenpos.common.exception.NotFoundException;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface TableGroupRepository extends JpaRepository<TableGroup, Long> {
+
+    default TableGroup tableGroup(long id) {
+        return findById(id).orElseThrow(
+            () -> new NotFoundException(String.format("단체 지정된 그룹 id(%d)를 찾을 수 없습니다.", id)));
+    }
+}
