+++ table/src/main/java/kichenpos/table/table/domain/CustomerStatus.java
@@ -1,0 +1,14 @@
+package kichenpos.table.table.domain;
+
+enum CustomerStatus {
+
+    EMPTY, SEATED, ORDERED, FINISH;
+
+    boolean isEmpty() {
+        return this == EMPTY;
+    }
+
+    boolean isOrdered() {
+        return this == ORDERED;
+    }
+}
