+++ table/src/main/java/kichenpos/table/table/ui/TableRestController.java
@@ -1,0 +1,58 @@
+package kichenpos.table.table.ui;
+
+import java.net.URI;
+import java.util.List;
+import kichenpos.table.table.application.TableService;
+import kichenpos.table.table.ui.request.OrderTableRequest;
+import kichenpos.table.table.ui.request.TableGuestsCountRequest;
+import kichenpos.table.table.ui.request.TableStatusRequest;
+import kichenpos.table.table.ui.response.OrderTableResponse;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.PutMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+@RequestMapping("/api/tables")
+public class TableRestController {
+
+    private final TableService tableService;
+
+    public TableRestController(TableService tableService) {
+        this.tableService = tableService;
+    }
+
+    @PostMapping
+    public ResponseEntity<OrderTableResponse> create(@RequestBody OrderTableRequest request) {
+        OrderTableResponse created = tableService.create(request);
+        URI uri = URI.create("/api/tables/" + created.getId());
+        return ResponseEntity.created(uri)
+            .body(created);
+    }
+
+    @GetMapping
+    public ResponseEntity<List<OrderTableResponse>> list() {
+        return ResponseEntity.ok()
+            .body(tableService.list());
+    }
+
+    @PutMapping("/{orderTableId}/empty")
+    public ResponseEntity<OrderTableResponse> changeEmpty(
+        @PathVariable long orderTableId,
+        @RequestBody TableStatusRequest request) {
+        return ResponseEntity.ok()
+            .body(tableService.changeEmpty(orderTableId, request));
+    }
+
+    @PutMapping("/{orderTableId}/number-of-guests")
+    public ResponseEntity<OrderTableResponse> changeNumberOfGuests(
+        @PathVariable long orderTableId,
+        @RequestBody TableGuestsCountRequest request) {
+        return ResponseEntity.ok()
+            .body(tableService.changeNumberOfGuests(orderTableId, request));
+    }
+}
