+++ table/src/main/resources/application.properties
@@ -1,0 +1,3 @@
+logging.level.org.springframework.jdbc.core=TRACE
+spring.h2.console.enabled=true
+spring.jpa.hibernate.ddl-auto=validate
