+++ table/src/main/resources/db/migration/V1__Initialize_project_tables.sql
@@ -1,0 +1,19 @@
+CREATE TABLE order_table
+(
+    id               BIGINT(20) NOT NULL AUTO_INCREMENT,
+    table_group_id   BIGINT(20),
+    number_of_guests INT(11) NOT NULL,
+    empty            BIT(1) NOT NULL,
+    PRIMARY KEY (id)
+);
+
+CREATE TABLE table_group
+(
+    id           BIGINT(20) NOT NULL AUTO_INCREMENT,
+    created_date DATETIME NOT NULL,
+    PRIMARY KEY (id)
+);
+
+ALTER TABLE order_table
+    ADD CONSTRAINT fk_order_table_table_group
+        FOREIGN KEY (table_group_id) REFERENCES table_group (id);
