+++ table/src/main/resources/db/migration/V2__Insert_default_data.sql
@@ -1,0 +1,8 @@
+INSERT INTO order_table (id, number_of_guests, empty) VALUES (1, 0, true);
+INSERT INTO order_table (id, number_of_guests, empty) VALUES (2, 0, true);
+INSERT INTO order_table (id, number_of_guests, empty) VALUES (3, 0, true);
+INSERT INTO order_table (id, number_of_guests, empty) VALUES (4, 0, true);
+INSERT INTO order_table (id, number_of_guests, empty) VALUES (5, 0, true);
+INSERT INTO order_table (id, number_of_guests, empty) VALUES (6, 0, true);
+INSERT INTO order_table (id, number_of_guests, empty) VALUES (7, 0, true);
+INSERT INTO order_table (id, number_of_guests, empty) VALUES (8, 0, true);
